requirements.txt
baidu-aip==4.16.1
beautifulsoup4==4.11.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
comtypes==1.1.11
fire==0.1.3
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
jieba==0.39
Jinja2==3.1.1
jsonpatch==1.32
jsonpointer==2.2
MarkupSafe==2.1.1
numpy==1.21.5
Pillow==9.0.1
pypiwin32==223
pyttsx==1.1
pyttsx3==2.90
pywin32==303
pyzmq==22.3.0
requests==2.27.1
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2
torch==1.11.0
torchfile==0.1.0
torchnet==0.0.4
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.8
visdom==0.1.8.9
websocket-client==1.3.1
Werkzeug==2.0.3
zipp==3.7.0
